rename: Header 컴포넌트 shared 폴더 이동 및 폴더명 camelCase 적용(#21)
